% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_match.R
\name{dist_match}
\alias{dist_match}
\title{Distribution Matching for Source and Reference Datasets}
\usage{
dist_match(
  src,
  ref,
  src_dist,
  ref_dist,
  lims,
  density = FALSE,
  samples = 1e+06,
  seed = NULL
)
}
\arguments{
\item{src}{Vector containing the source data to be matched.}

\item{ref}{Vector containing the reference data to estimate the reference
distribution for matching.}

\item{src_dist}{Estimation of the source distribution. If missing, it is
estimated from the provided source data.}

\item{ref_dist}{Estimation of the reference distribution. If missing, it is
estimated from the provided reference data.}

\item{lims}{Vector giving the range for the knot values. If missing, these
values are estimated from the given source data.}

\item{density}{Flag for using kernel density estimates for matching
distributions instead of histogram counts. Defaults to \code{False}.}

\item{samples}{Sample size for estimating distributions if \code{src_dist} and/or
\code{ref_dist} are missing. Defaults to \code{1e6}.}

\item{seed}{Seed for random number generator (for reproducible outcomes).
Defaults to \code{NULL}.}
}
\description{
This function matches the source distribution to a reference distribution
so that the source data can effectively transferred to the reference space.
}
\examples{

}
\keyword{distribution-matching}
\keyword{domain-transfer}
