% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_est.R
\name{get_dist_est}
\alias{get_dist_est}
\title{Estimate cumulative distribution}
\usage{
get_dist_est(
  x,
  size_tol = 1000,
  sample_size = 1e+06,
  density = FALSE,
  binned = TRUE,
  grid_size = 10000,
  x_range = "est",
  random_seed = NULL,
  ...
)
}
\arguments{
\item{x}{Vector containing data.}

\item{size_tol}{The size tolerance for performing bootstrapping on \code{x} to
estimate distribution values. Defaults to \code{1e3}.}

\item{sample_size}{Sample size used for bootstrapping. Defaults to \code{1e6}.}

\item{density}{Flag for calculating kernel density estimates instead of
histogram counts. Uses functions from the \code{ks} package to estimate kernel
density values. Defaults to \code{FALSE}.}

\item{binned}{Flag for calculating binned estimates of kernel density if
\code{density = TRUE}.}

\item{grid_size}{Size of estimation grid. Defaults to \code{1e4}.}

\item{x_range}{Range of data values. Options are either \code{"unit"} (normalized
in \link{0, 1}) or \code{"est"} (estimated from data). Defaults to \code{"est"}.}

\item{random_seed}{Seed for random number generator (for reproducibility).
Defaults to \code{NULL}.}

\item{...}{Other options relevant for distribution estimation.}
}
\description{
This function estimates the values of the cumulative distribution function
(CDF) for a vector.
}
\examples{
x <- runif(100)
x_cdf <- get_dist_est(x, density = FALSE, x_range = "unit")

}
\keyword{CDF}
\keyword{cumulative-distribution}
