% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_functions.R
\name{norm01}
\alias{norm01}
\title{Normalize vector in [0, 1]}
\usage{
norm01(x)
}
\arguments{
\item{x}{Vector containing data.}
}
\value{
The normalized vector.
}
\description{
This function normalizes a given vector between 0 and 1.
}
\examples{
x <- rnorm(100, 0.2, 0.3)
x_norm <- norm01(x)
print(range(x_norm))

}
\keyword{normalization}
