% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_modeling.R
\name{RF_predict}
\alias{RF_predict}
\title{Predictive Modeling using Random Forest Regression}
\usage{
RF_predict(
  x_train,
  y_train,
  x_test,
  lims,
  n_tree = 300,
  m_try = 0.3333,
  seed = NULL,
  ...
)
}
\arguments{
\item{x_train}{Training features for designing the RF regressor.}

\item{y_train}{Training response for designing the RF regressor.}

\item{x_test}{Test features for which response values are to be predicted.}

\item{lims}{Vector providing the range of the response values for modeling.
If missing, these values are estimated from the training response.}

\item{n_tree}{Number of decision trees to be built in the forest. Defaults
to \code{300}.}

\item{m_try}{Fraction of the features to be used for building each tree.
Defaults to 0.3333 (or 33.33\%).}

\item{seed}{Seed for random number generator (for reproducible outcomes).
Defaults to \code{NULL}.}

\item{...}{Other options relevant for RF modeling.}
}
\description{
This function trains a Random Forest regressor using the training data
provided and predict response for the test features. Depends on the
\code{randomForest} package.
}
\examples{

}
\keyword{decision-tree}
\keyword{random-forest}
