% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_match_trans_learn.R
\name{DMTL}
\alias{DMTL}
\title{Distribution Matching based Transfer Learning}
\usage{
DMTL(
  target_set,
  source_set,
  use_density = FALSE,
  sample_size = 1000,
  random_seed = NULL,
  all_pred = FALSE
)
}
\arguments{
\item{target_set}{List containing the target datasets. A named list with
components \code{X} (features) and \code{y} (response). The predictions are performed
to estimate the response values corresponding to \code{X} while \code{y} is only used
to estimate the response distribution parameters.}

\item{source_set}{List containing the source datasets. A named list with
components \code{X} (features) and \code{y} (response). These two sets must be matched
and used in both distribution estimation and predictive modeling.}

\item{use_density}{Flag for using kernel density as distribution estimate
instead of histogram counts. Defaults to \code{FALSE}.}

\item{sample_size}{Sample size for estimating distributions of target and
source datasets. Defaults to \code{1e3}.}

\item{random_seed}{Seed for random number generator (for reproducible
outcomes). Defaults to \code{NULL}.}

\item{all_pred}{Flag for returning the prediction values in the source space.
If \code{TRUE}, \code{DMTL()} returns a named list with two components- \code{target} and
\code{source} (predictions in the target space and source space, respectively).
Defaults to \code{FALSE}.}
}
\value{
If \code{all_pred = FALSE}, a vector containing the final predictions.

If \code{all_pred = TRUE}, a named list with two components \code{target} and \code{source}
\emph{i.e.}, predictions in the original target space and in source space,
respectively.
}
\description{
This function performs distribution matching based transfer learning (DMTL)
regression for given target (primary) and source (secondary) datasets. The
data available in the source domain are used to design an appropriate
predictive model. The target features with unknown response values are
transferred to the source domain \emph{via} distribution matching and then the
corresponding response values in the source domain are predicted using the
aforementioned predictive model. The response values are then transferred to
the original target space by applying distribution matching again. Hence,
this function needs an \strong{unmatched} pair of target datasets (features and
response values) and a \strong{matched} pair of source datasets.
}
\note{
The datasets in \code{target_set} (\emph{i.e.}, \code{X} and \code{y}) do not need to be
matched (\emph{i.e.}, have the same number of rows) since the response values are
used only to estimate distribution for mapping while the feature values are
used for both mapping and final prediction. In contrast, the datasets in
\code{source_set} (\emph{i.e.}, \code{X} and \code{y}) must have matched samples.
}
\examples{

}
\keyword{distribution-matching}
\keyword{domain-transfer}
\keyword{transfer-learning}
