% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_match_trans_learn.R
\name{DMTL}
\alias{DMTL}
\title{Distribution Matching based Transfer Learning}
\usage{
DMTL(
  target_set,
  source_set,
  use_density = FALSE,
  sample_size = 1000,
  random_seed = NULL,
  all_pred = FALSE
)
}
\arguments{
\item{target_set}{List containing the target datasets. A named list with two
components- \code{x} (features) and \code{y} (response). These two sets do not need to
be matched and the response values are only used to estimate distribution
while the corresponding response values for the features are predicted.}

\item{source_set}{List containing the source datasets. A named list with two
components- \code{x} (features) and \code{y} (response). These two sets must be matched
and used in both distribution estimation and training a predictive model.}

\item{use_density}{Flag for using kernel density as distribution estimate
instead of histogram counts. Defaults to \code{FALSE}.}

\item{sample_size}{Sample size for estimating distributions of target and
source datasets. Defaults to \code{1e3}.}

\item{random_seed}{Seed for random number generator (for reproducible
outcomes). Defaults to \code{NULL}.}

\item{all_pred}{Flag for returning the prediction values in the source space.
If \code{TRUE}, \code{DMTL()} returns a named list with two components- \code{target} and
\code{source} (predictions in the target space and source space, respectively).
Defaults to \code{FALSE}.}
}
\description{
This function performs distribution matching based transfer learning (DMTL)
regression for given target (primary) and source (secondary) datasets. The
data available in the source domain are used to design an appropriate
predictive model. The target features with unknown response values are
transferred to the source domain \emph{via} distribution matching and then the
corresponding response values in the source domain are predicted using the
aforementioned predictive model. The response values are then transferred to
the original target space by applying distribution matching again. Hence,
this function needs an \strong{unmatched} pair of target datasets (features and
response values) and a \strong{matched} pair of source datasets.
}
\examples{

}
\keyword{distribution-matching}
\keyword{domain-transfer}
\keyword{transfer-learning}
