% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_match.R
\name{match_func}
\alias{match_func}
\title{Estimate inverse function mapping}
\usage{
match_func(knots, vals, new_vals, lims)
}
\arguments{
\item{knots}{Vector containing knots for the distribution estimate.}

\item{vals}{Vector containing distribution values corresponding to the knots.}

\item{new_vals}{Vector containing distribution values for which the knots
are unknown.}

\item{lims}{Vector giving the range for the knot values. If missing, these
values are estimated from the given knots.}
}
\description{
This function estimates an inverse map for a distribution estimate from its
knots and value sets, and calculates corresponding knots for new values
assumed to be from the distribution.
}
\keyword{inverse-map}
