% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_functions.R
\name{norm_data}
\alias{norm_data}
\title{Normalize matrix per column in [0, 1]}
\usage{
norm_data(X)
}
\arguments{
\item{X}{Dataframe or matrix (-alike) containing data.}
}
\value{
The normalized dataframe.
}
\description{
This function normalizes each column of a dataframe or matrix (-alike)
between 0 and 1.
}
\examples{
X <- matrix(rnorm(1000, 0.2, 0.3), nrow = 100)
X_norm <- norm_data(X)
print(range(X_norm))

}
\keyword{normalization}
